Proxy DependencyMap interface
This interface implements a Proxy pattern instead of the Descriptor
one. The idea is give to the developer the power of the di-py library
but hidden it to build class global variables.

A new ProxyDepencyMap class is able to generate InjectorProxy to act
as a Proxy of the proper instance given by the DependencyMap.

For example:

    >>> from di import ProxyDependencyMap
    >>> from .services import MyService
    >>>
    >>> _dm = ProxyDependencyMap()
    >>>
    >>> @_dm.singleton(MyService):
    >>> def _build_my_service(deps):
    >>>     return MyService()
    >>>
    >>> my_service = _dm(MyService)